early stopping added

+++ src/torch_dataset/sequantial_image.py
@@ -8,7 +8,8 @@ class SequentialImage2Classify(Dataset):
     def __init__(self, dataset, device_name=None):
         self.data = []
         for image, label in dataset:
-            image = image.to(dtype=torch.float32).view(image.shape[0], -1)
+            image = image.to(dtype=torch.float32)
+            image = image.view(image.shape[0], -1)
             image = normalize(image, p=2, dim=-1)
             label = torch.tensor(label, dtype=torch.long)
             if device_name is not None:
